refactor(web): routes配下の機能hook・コンポーネントをfeaturesへ移動
routes/ にはルート組み立て（route/page/layout）のみを残し、機能（ドメイン）の
hook・コンポーネントを各 feature スライスへ移し、スライス外からは Public API
（index.ts）経由で参照する形へ揃える。

- authenticate: use-login / use-signup を features/authenticate/hooks へ
- diary: use-diary / use-analytics / use-diary-api を features/diary/hooks へ
- article: trends 配下のコンポーネント・hook を新設 features/article へ集約し、
  media-icon もここへ移動（diary からは Public API 経由で参照）
- inbox: use-unread-digestion を新設 features/inbox へ
- analytics-page.tsx は利用元ルートの routes/analytics/page.tsx へ移し、
  page.test.ts を diary / analytics に分割
