+++ application/packages/web/src/client/features/article/components/media-icon.tsx
