+++ application/packages/web/src/client/features/authenticate/hooks/use-signup.ts
@@ -1,19 +1,18 @@
 import { wrapAsyncCall } from '@trend-diary/common/result'
 import { useState } from 'react'
 import { useNavigate } from 'react-router'
-import { useSWRConfig } from 'swr'
 import {
   AUTH_ERROR_MESSAGES,
+  resolveSignupErrorMessage,
+} from '@/client/features/authenticate/error-message'
+import {
   type AuthenticateErrors,
-  resolveLoginErrorMessage,
-  SESSION_SWR_KEY,
   validateAuthenticateForm,
-} from '@/client/features/authenticate'
+} from '@/client/features/authenticate/validation'
 import getApiClientForClient from '@/client/infrastructure/api'
 
-export default function useLogin(turnstileSiteKey?: string) {
+export default function useSignup(turnstileSiteKey?: string) {
   const navigate = useNavigate()
-  const { mutate } = useSWRConfig()
   const [isSubmitting, setIsSubmitting] = useState(false)
   const [errors, setErrors] = useState<AuthenticateErrors | undefined>(undefined)
   const [formError, setFormError] = useState<string | undefined>(undefined)
@@ -39,7 +38,7 @@ export default function useLogin(turnstileSiteKey?: string) {
     setIsSubmitting(true)
     const result = await wrapAsyncCall(() => {
       const client = getApiClientForClient()
-      return client.auth.login.$post({
+      return client.auth.signup.$post({
         json: {
           email: validation.data.email,
           password: validation.data.password,
@@ -54,15 +53,13 @@ export default function useLogin(turnstileSiteKey?: string) {
       return
     }
     if (!result.value.ok) {
-      setFormError(resolveLoginErrorMessage(result.value.status))
+      setFormError(resolveSignupErrorMessage(result.value.status))
       setIsSubmitting(false)
       return
     }
 
-    // ログイン前の未ログイン状態がセッションキャッシュに残ったまま遷移しないよう再検証する
-    await mutate(SESSION_SWR_KEY)
     // 成功時は遷移でアンマウントされるため、ボタンを無効のままにして二重送信を防ぐ
-    navigate('/trends')
+    navigate('/login')
   }
 
   return { isSubmitting, errors, formError, submit }
