+++ application/packages/web/src/client/features/authenticate/index.ts
@@ -5,7 +5,9 @@ export {
   resolveSignupErrorMessage,
 } from './error-message'
 export { resolveTurnstileSiteKey } from './turnstile'
+export { default as useLogin } from './hooks/use-login'
 export { default as useSession, SESSION_SWR_KEY } from './hooks/use-session'
+export { default as useSignup } from './hooks/use-signup'
 export {
   type AuthenticateErrors,
   type AuthenticateFormData,
