+++ application/packages/web/src/client/features/diary/index.ts
@@ -4,4 +4,6 @@ export { default as DiaryReadListSection } from './components/read-list-section'
 export { default as DiaryReadPagination } from './components/read-pagination'
 export { default as DiarySummarySection } from './components/summary-section'
 export { getTodayJst, sumSourceSummary } from './daily-summary'
+export { default as useAnalytics } from './hooks/use-analytics'
+export { default as useDiary } from './hooks/use-diary'
 export type { ReadItem, ReadPagination, Source, Summary } from './types'
