+++ application/packages/web/src/client/features/inbox/hooks/use-unread-digestion.test.ts
@@ -1,17 +1,16 @@
 import { act, renderHook, waitFor } from '@testing-library/react'
 import { createElement, type ReactNode } from 'react'
 import { SWRConfig } from 'swr'
+import { type MediaType, useReadArticle } from '@/client/features/article'
 import createSWRFetcher from '@/client/infrastructure/create-swr-fetcher'
-import type { MediaType } from '../../trends._index/components/media-filter'
-import useReadArticle from '../../trends._index/hooks/use-read-article'
 import useUnreadDigestion, { type Article } from './use-unread-digestion'
 
 vi.mock('@/client/infrastructure/create-swr-fetcher', () => ({
   default: vi.fn(),
 }))
 
-vi.mock('../../trends._index/hooks/use-read-article', () => ({
-  default: vi.fn(),
+vi.mock('@/client/features/article', () => ({
+  useReadArticle: vi.fn(),
 }))
 
 const mockedCreateSWRFetcher = vi.mocked(createSWRFetcher)
