+++ application/packages/web/src/client/features/inbox/hooks/use-unread-digestion.ts
@@ -2,9 +2,8 @@ import type { ArticleOutput } from '@trend-diary/domain/article/schema/article-s
 import { useEffect, useRef, useState } from 'react'
 import { toast } from 'sonner'
 import useSWR from 'swr'
+import { type MediaType, useReadArticle } from '@/client/features/article'
 import createSWRFetcher from '@/client/infrastructure/create-swr-fetcher'
-import type { MediaType } from '../../trends._index/components/media-filter'
-import useReadArticle from '../../trends._index/hooks/use-read-article'
 
 export type Article = Omit<ArticleOutput, 'articleId'> & {
   articleId: string
