+++ application/packages/web/src/client/features/inbox/index.ts
@@ -1,0 +1,1 @@
+export { default as useUnreadDigestion, type Article } from './hooks/use-unread-digestion'
