+++ application/packages/web/src/client/routes/analytics/page.tsx
