+++ application/packages/web/src/client/routes/trends._index/page.tsx
@@ -11,12 +11,17 @@ import {
   PaginationNext,
   PaginationPrevious,
 } from '@/client/components/shadcn/pagination'
+import {
+  type Article,
+  ArticleCard,
+  DatePresetFilter,
+  type DatePresetType,
+  MediaFilter,
+  type MediaType,
+  ReadStatusFilter,
+  type ReadStatusType,
+} from '@/client/features/article'
 import LoadingSpinner from '../../components/ui/feedback/loading-spinner'
-import ArticleCard from './components/article-card'
-import DatePresetFilter from './components/date-preset-filter'
-import MediaFilter, { type MediaType } from './components/media-filter'
-import ReadStatusFilter, { type ReadStatusType } from './components/read-status-filter'
-import type { Article, DatePresetType } from './hooks/use-articles'
 
 const scrollToTop = () => {
   window.scrollTo({ top: 0, behavior: 'smooth' })
